refactor: split snapshot projection and operator-stop latch out of orchestrator state.go (#661)
Pure behavior-preserving move: the SPEC §13.3/§13.7 snapshot projection
(StateView, snapshotRunningViews, Snapshot, copyRateLimitSnapshot) moves
to state_snapshot.go next to the views.go projection row types, and the
D35 operator-terminal-stop latch (OperatorTerminalStopEntry plus its
record/query/suppression-accounting methods) moves to
state_operator_stop.go, leaving state.go at 716 lines — under the
800-line production-file budget. The file's entry is removed from the
file-size baseline so the ratchet now enforces the budget for it
directly.
