+++ internal/orchestrator/state.go
@@ -12,7 +12,6 @@ package orchestrator
 
 import (
 	"context"
-	"strings"
 	"time"
 
 	"github.com/xrf9268-hue/aiops-platform/internal/tracker"
@@ -196,21 +195,6 @@ type BlockedEntry struct {
 	Error       string
 }
 
-// OperatorTerminalStopEntry is a process-local latch recording that this
-// worker observed an operator move an issue to a terminal tracker state. D35
-// makes that observation authoritative for this worker process so a later
-// agent-side issueUpdate cannot re-activate and re-dispatch the same issue.
-type OperatorTerminalStopEntry struct {
-	Issue                 tracker.Issue
-	Identifier            string
-	State                 string
-	StoppedAt             time.Time
-	SuppressedDispatches  int
-	FirstSuppressedAt     time.Time
-	FirstSuppressedState  string
-	FirstSuppressedReason string
-}
-
 // OrchestratorState is the single authoritative in-memory state owned
 // by the orchestrator (SPEC §4.1.8). Every field name matches the SPEC
 // section exactly.
@@ -383,75 +367,6 @@ func (s *OrchestratorState) MarkCleaningWorkspace(id IssueID) {
 // UnmarkCleaningWorkspace clears the cleanup-in-flight mark for id.
 func (s *OrchestratorState) UnmarkCleaningWorkspace(id IssueID) {
 	delete(s.cleaningWorkspaces, id)
-}
-
-func (s *OrchestratorState) RecordOperatorTerminalStop(id IssueID, issue tracker.Issue, at time.Time) (OperatorTerminalStopEntry, bool) {
-	if s.OperatorTerminalStops == nil {
-		s.OperatorTerminalStops = map[IssueID]*OperatorTerminalStopEntry{}
-	}
-	if at.IsZero() {
-		at = time.Now().UTC()
-	}
-	identifier := issue.Identifier
-	if identifier == "" {
-		identifier = string(id)
-	}
-	state := issue.State
-	if existing, ok := s.OperatorTerminalStops[id]; ok {
-		if strings.TrimSpace(existing.Identifier) == "" {
-			existing.Identifier = identifier
-		}
-		if strings.TrimSpace(state) != "" {
-			existing.State = state
-			existing.Issue = issue
-		}
-		return *existing, false
-	}
-	entry := &OperatorTerminalStopEntry{
-		Issue:      issue,
-		Identifier: identifier,
-		State:      state,
-		StoppedAt:  at,
-	}
-	s.OperatorTerminalStops[id] = entry
-	s.operatorTerminalStopOrder = append(s.operatorTerminalStopOrder, id)
-	// Counted here, after the repeat early-return — intentionally unlike
-	// recordCompleted, which counts every call. A re-observed stop is not a new stop.
-	s.CumulativeOperatorTerminalStopsTotal++
-	if s.MaxRecentOperatorTerminalStops > 0 && len(s.operatorTerminalStopOrder) > s.MaxRecentOperatorTerminalStops {
-		oldest := s.operatorTerminalStopOrder[0]
-		s.operatorTerminalStopOrder = s.operatorTerminalStopOrder[1:]
-		delete(s.OperatorTerminalStops, oldest)
-	}
-	return *entry, true
-}
-
-func (s *OrchestratorState) IsOperatorTerminalStopped(id IssueID) bool {
-	_, ok := s.OperatorTerminalStops[id]
-	return ok
-}
-
-func (s *OrchestratorState) LookupOperatorTerminalStop(id IssueID) (OperatorTerminalStopEntry, bool) {
-	entry, ok := s.OperatorTerminalStops[id]
-	if !ok || entry == nil {
-		return OperatorTerminalStopEntry{}, false
-	}
-	return *entry, true
-}
-
-func (s *OrchestratorState) RecordOperatorTerminalDispatchSuppressed(id IssueID, issue tracker.Issue, reason string) (OperatorTerminalStopEntry, bool) {
-	entry, ok := s.OperatorTerminalStops[id]
-	if !ok || entry == nil {
-		return OperatorTerminalStopEntry{}, false
-	}
-	entry.SuppressedDispatches++
-	first := entry.FirstSuppressedAt.IsZero()
-	if first {
-		entry.FirstSuppressedAt = time.Now().UTC()
-		entry.FirstSuppressedState = issue.State
-		entry.FirstSuppressedReason = reason
-	}
-	return *entry, first
 }
 
 // IsClaimed reports whether id is currently held by any of Running,
@@ -798,205 +713,4 @@ func (s *OrchestratorState) RecordCodexTokens(input, output int64) {
 // view, not a delta.
 func (s *OrchestratorState) RecordRateLimits(snap *RateLimitSnapshot) {
 	s.CodexRateLimits = copyRateLimitSnapshot(snap)
-}
-
-// StateView is the SPEC §13.3 / §13.7 shape the orchestrator publishes
-// for /api/v1/state, CLI status, and tests. It is intentionally a
-// value type with slices so callers cannot mutate the orchestrator's
-// internal maps through it. The HTTP wrapper is deferred (see the
-// design doc's "HTTP surface" section) but Snapshot() ships now so
-// tests have a stable read API.
-type StateView struct {
-	GeneratedAt                time.Time
-	PollIntervalMs             int64
-	MaxConcurrentAgents        int
-	MaxConcurrentAgentsByState map[string]int
-
-	Running  []RunningView
-	Blocked  []BlockedView
-	Retrying []RetryView
-	// Completed is bounded by MaxRecentCompleted on the source state; the
-	// slice here is FIFO-ordered (oldest first). For the lifetime total that
-	// survives eviction, read CumulativeCompletedTotal.
-	Completed []IssueID
-	// ReconcileStoppedWithProgress is the FIFO-bounded recent set of reconcile-stopped
-	// runs that had completed ≥1 agent turn (made progress — usually the agent's
-	// handoff, but inspect to confirm) — surfaced so a progressed-but-reaped run is
-	// visible rather than absent from Completed (#557).
-	// CumulativeReconcileStoppedWithProgressTotal is the lifetime monotonic counter
-	// that survives FIFO eviction.
-	ReconcileStoppedWithProgress []IssueID
-	// AgentHandoffReconcileStopped is the FIFO-bounded recent set of
-	// reconcile-stopped runs that observed a guarded current-issue Linear state
-	// handoff. It may overlap ReconcileStoppedWithProgress when the handoff also
-	// completed a turn before reconcile reaped it.
-	AgentHandoffReconcileStopped []IssueID
-	// OperatorTerminalStops is the FIFO-bounded recent set of D35 latches (oldest
-	// first), capped by MaxRecentOperatorTerminalStops. For the lifetime total that
-	// survives eviction, read CumulativeOperatorTerminalStopsTotal.
-	OperatorTerminalStops                       []OperatorTerminalStopView
-	CumulativeCompletedTotal                    int64
-	CumulativeReconcileStoppedWithProgressTotal int64
-	CumulativeAgentHandoffReconcileStoppedTotal int64
-	CumulativeOperatorTerminalStopsTotal        int64
-	CodexTotals                                 CodexTotals
-	CodexRateLimits                             *RateLimitSnapshot
-	// RecentEvents is the bounded orchestrator-wide event log (capped at
-	// MaxRuntimeEvents). It carries SPEC §13.7 operator-visible signals
-	// like dispatch_preflight_failed that are not scoped to one issue.
-	RecentEvents []RuntimeEvent
-}
-
-// The StateView projection types (RunningView, BlockedView, RetryView,
-// TokensView, OperatorTerminalStopView) live in views.go; their builders stay
-// here next to Snapshot.
-
-// Snapshot returns a read-only view of the orchestrator state. The
-// returned slices are freshly allocated so the caller may sort or
-// truncate them without affecting future snapshots. Map iteration
-// order is unspecified in Go, but downstream consumers (the §13.7 HTTP
-// handler, CLI status) sort by IssueID before display.
-// snapshotRunningViews projects s.Running into RunningView rows. It deep-copies
-// the RetryAttempt pointer so a snapshot consumer mutating the pointee cannot
-// reach back into orchestrator state; the pointer-vs-nil distinction is the SPEC
-// §4.1.5 first-run semantic, so it cannot be flattened to an int.
-func (s *OrchestratorState) snapshotRunningViews() []RunningView {
-	rows := make([]RunningView, 0, len(s.Running))
-	for id, r := range s.Running {
-		var retryAttempt *int
-		if r.RetryAttempt != nil {
-			n := *r.RetryAttempt
-			retryAttempt = &n
-		}
-		rows = append(rows, RunningView{
-			IssueID:       id,
-			Identifier:    r.Identifier,
-			IssueURL:      r.Issue.URL,
-			State:         r.Issue.State,
-			SessionID:     r.Session.SessionID,
-			TurnCount:     r.Session.TurnCount,
-			LastEvent:     r.LastCodexEvent,
-			LastMessage:   r.LastCodexMessage,
-			StartedAt:     r.StartedAt,
-			LastEventAt:   r.LastEventAt,
-			RetryAttempt:  retryAttempt,
-			WorkspacePath: r.Workspace.Path,
-			Tokens: TokensView{
-				InputTokens:  r.CodexInputTokens,
-				OutputTokens: r.CodexOutputTokens,
-				TotalTokens:  r.CodexTotalTokens,
-			},
-			CodexAppServerPID: r.Session.CodexAppServerPID,
-		})
-	}
-	return rows
-}
-
-func (s *OrchestratorState) Snapshot() StateView {
-	// Keep the live-aggregate math on the monotonic clock: `time.Now()`
-	// carries a monotonic component that `time.Time.Sub` uses for elapsed
-	// calculations, but `time.Time.UTC()` strips it (Go stdlib documents
-	// the loss explicitly). RunningEntry.StartedAt is set from `time.Now()`
-	// in spawn (with monotonic reading intact), so subtracting two
-	// monotonic-bearing times keeps the elapsed reading invariant under
-	// wall-clock NTP steps. `GeneratedAt` is the operator-visible
-	// timestamp and needs UTC; do the strip only when populating it.
-	now := time.Now()
-	// SPEC §13.5 Runtime accounting: report runtime as a live aggregate at
-	// snapshot/render time. CodexTotals.SecondsRunning increments only on
-	// session end (AddSeconds in the worker-exit finalization path); without
-	// the per-snapshot fold a dashboard polling /api/v1/state would see a
-	// flat counter while a long run is in flight and a sudden jump on exit.
-	// Add elapsed time for each currently-running entry, measured against
-	// the same monotonic `now` so all rows share one instant. The
-	// ended-session counter and the live aggregate cannot double-count
-	// because a finished run is removed from s.Running before AddSeconds
-	// runs for it.
-	totals := s.CodexTotals
-	for _, r := range s.Running {
-		if r.StartedAt.IsZero() {
-			continue
-		}
-		if elapsed := now.Sub(r.StartedAt); elapsed > 0 {
-			totals.SecondsRunning += elapsed.Seconds()
-		}
-	}
-	view := StateView{
-		GeneratedAt:                  now.UTC(),
-		PollIntervalMs:               s.PollIntervalMs,
-		MaxConcurrentAgents:          s.MaxConcurrentAgents,
-		MaxConcurrentAgentsByState:   copyStateConcurrencyLimits(s.MaxConcurrentAgentsByState),
-		Blocked:                      make([]BlockedView, 0, len(s.Blocked)),
-		Retrying:                     make([]RetryView, 0, len(s.RetryAttempts)),
-		Completed:                    make([]IssueID, 0, len(s.completedOrder)),
-		ReconcileStoppedWithProgress: make([]IssueID, 0, len(s.reconcileStoppedWithProgressOrder)),
-		AgentHandoffReconcileStopped: make([]IssueID, 0, len(s.agentHandoffReconcileStoppedOrder)),
-		OperatorTerminalStops:        make([]OperatorTerminalStopView, 0, len(s.operatorTerminalStopOrder)),
-		CumulativeCompletedTotal:     s.CumulativeCompletedTotal,
-		CumulativeReconcileStoppedWithProgressTotal: s.CumulativeReconcileStoppedWithProgressTotal,
-		CumulativeAgentHandoffReconcileStoppedTotal: s.CumulativeAgentHandoffReconcileStoppedTotal,
-		CumulativeOperatorTerminalStopsTotal:        s.CumulativeOperatorTerminalStopsTotal,
-		CodexTotals:                                 totals,
-		CodexRateLimits:                             copyRateLimitSnapshot(s.CodexRateLimits),
-		RecentEvents:                                append([]RuntimeEvent(nil), s.RecentEvents...),
-	}
-	view.Running = s.snapshotRunningViews()
-	for id, b := range s.Blocked {
-		view.Blocked = append(view.Blocked, BlockedView{
-			IssueID:           id,
-			Identifier:        b.Identifier,
-			IssueURL:          b.Issue.URL,
-			State:             b.Issue.State,
-			BlockedAt:         b.BlockedAt,
-			WorkspacePath:     b.Workspace.Path,
-			SessionID:         b.Session.SessionID,
-			LastEventAt:       b.LastEventAt,
-			Method:            b.Method,
-			Error:             b.Error,
-			CodexAppServerPID: b.Session.CodexAppServerPID,
-		})
-	}
-	for id, r := range s.RetryAttempts {
-		view.Retrying = append(view.Retrying, RetryView{
-			IssueID:    id,
-			Identifier: r.Identifier,
-			IssueURL:   r.Issue.URL,
-			Attempt:    r.Attempt,
-			DueAt:      r.DueAt,
-			Error:      r.Error,
-			Kind:       r.Kind,
-		})
-	}
-	// Iterate the FIFO order slices, not the map. The map's iteration
-	// order is undefined in Go; using the slices preserves observed
-	// insertion order so /api/v1/state consumers see "oldest first"
-	// stably, and the bounded slice matches MaxRecent* exactly.
-	view.Completed = append(view.Completed, s.completedOrder...)
-	view.ReconcileStoppedWithProgress = append(view.ReconcileStoppedWithProgress, s.reconcileStoppedWithProgressOrder...)
-	view.AgentHandoffReconcileStopped = append(view.AgentHandoffReconcileStopped, s.agentHandoffReconcileStoppedOrder...)
-	for _, id := range s.operatorTerminalStopOrder {
-		entry, ok := s.LookupOperatorTerminalStop(id)
-		if !ok {
-			continue
-		}
-		view.OperatorTerminalStops = append(view.OperatorTerminalStops, OperatorTerminalStopView{
-			IssueID:               id,
-			Identifier:            entry.Identifier,
-			State:                 entry.State,
-			StoppedAt:             entry.StoppedAt,
-			SuppressedDispatches:  entry.SuppressedDispatches,
-			FirstSuppressedAt:     entry.FirstSuppressedAt,
-			FirstSuppressedState:  entry.FirstSuppressedState,
-			FirstSuppressedReason: entry.FirstSuppressedReason,
-		})
-	}
-	return view
-}
-
-func copyRateLimitSnapshot(in *RateLimitSnapshot) *RateLimitSnapshot {
-	if in == nil {
-		return nil
-	}
-	copied := RateLimitSnapshot(copyAnyMap(map[string]any(*in)))
-	return &copied
 }
